Added texture foreground, font_fallback, font thickness/border feature
