+++ test/utf8_sample.txt
@@ -1,1 +1,1 @@
-﻿Μπορώ να φάω σπασμένα γυαλιά χωρίς να πάθω τίποτα.
+﻿Μπορώ να φάω σπασμένα γυαλιά χωρίς να πάθω τίποτα.아에이오우あえいおうabcde
